[0.11.3] add api base helper
