+++ src/app/api/v1/auth/create.rb
@@ -24,19 +24,6 @@ module API
             { ip: request.ip,
               fingerprint: params['device']['fingerprint'] }
           end
-
-          # TODO: move to global helpers
-          def authenticator_error(status)
-            type = case status.to_sym
-                   when :device_verification_needed
-                     API::Errors::DeviceNotVerifiedError
-                   when :invalid_email_or_password
-                     API::Errors::InvalidEmailOrPasswordError
-                   else
-                     API::Errors::UnexpectedError
-                   end
-            type.new(status)
-          end
         end
 
         version :v1 do
